requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic-settings
python-dotenv
supabase
passlib[bcrypt]
python-jose[cryptography]
pydantic-ai
pydantic[email]
